# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Agora CMS
  description: Central management system for Agora media playback devices
  version: 1.38.249
paths:
...
  /api/notifications:
    get:
      summary: List Notifications
      description: List notifications visible to the current user, newest first.
      operationId: list_notifications_api_notifications_get
      parameters:
      - name: unread_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Unread Only
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationOut'
                title: Response List Notifications Api Notifications Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete All Notifications
      description: Dismiss all visible notifications for the current user.
      operationId: delete_all_notifications_api_notifications_delete
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
...
  /api/notifications/{notification_id}/read:
    post:
      summary: Mark Read
      description: Mark a single notification as read for the current user.
      operationId: mark_read_api_notifications__notification_id__read_post
      parameters:
      - name: notification_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Notification Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/notifications/{notification_id}:
    delete:
      summary: Delete Notification
      description: Dismiss a notification for the current user.
      operationId: delete_notification_api_notifications__notification_id__delete
      parameters:
      - name: notification_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Notification Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NotificationOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        scope:
          type: string
          title: Scope
        level:
          type: string
          title: Level
        title:
          type: string
          title: Title
        message:
          type: string
          title: Message
        details:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Details
        group_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Group Id
        group_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Group Ids
          default: []
        user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: User Id
        created_at:
          type: string
          format: date-time
          title: Created At
        read_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Read At
      type: object
      required:
      - id
      - scope
      - level
      - title
      - message
      - created_at
      title: NotificationOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Browser session cookie issued by POST /login.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Static API key issued via /api/keys; used for device-originated and automation callers.
servers:
- url: http://localhost:8080
  description: Local Docker Compose
security:
- sessionCookie: []
...